# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-proxy-api",
        "contact": {},
        "version": "0.1.2"
    },
    "basePath": "/api",
    "paths": {
        "/": {
            "get": {
                "description": "Ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Ping",
                "operationId": "/.get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routers.PingOutputScheme"
                        }
                    }
                }
            }
        },
        "/apple/location/{code}": {
            "get": {
                "description": "返回 code 文本",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Apple"
                ],
                "summary": "Location Code",
                "operationId": "apple.location.code",
                "parameters": [
                    {
                        "type": "string",
                        "example": "US",
                        "description": "地区代码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/convert/svg/png": {
            "get": {
                "description": "将 SVG 图片转为 png",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "convert"
                ],
                "summary": "SVG to PNG",
                "operationId": "convert.svg.png.get",
                "parameters": [
                    {
                        "type": "string",
                        "example": "https://www.docker.com/wp-content/uploads/2024/01/icon-docker-square.svg",
                        "description": "svg图片地址",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "是否下载",
                        "name": "download",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/convert/xml/json": {
            "get": {
                "description": "将传入的 xml 字符串转成 json字符串并返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "convert"
                ],
                "summary": "XML to JSON",
                "operationId": "convert.xml.json.get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "xml字符串",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/xml.ConvertXMLOutput"
                        }
                    }
                }
            },
            "post": {
                "description": "将传入的 xml 字符串转成 json字符串并返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "convert"
                ],
                "summary": "XML to JSON",
                "operationId": "convert.xml.json.post",
                "parameters": [
                    {
                        "description": "-",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/xml.ConvertXMLInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/xml.ConvertXMLOutput"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Ping",
                "operationId": "ping.get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routers.PingOutputScheme"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "routers.PingDocsScheme": {
            "type": "object",
            "required": [
                "docs",
                "redoc"
            ],
            "properties": {
                "docs": {
                    "description": "指向 Swagger UI 文档格式的页面",
                    "type": "string",
                    "example": "/docs/index.html"
                },
                "redoc": {
                    "description": "指向 Redoc 文档格式的页面",
                    "type": "string",
                    "example": "/redoc"
                }
            }
        },
        "routers.PingOutputScheme": {
            "type": "object",
            "required": [
                "current",
                "docs",
                "message",
                "run_at",
                "run_at_ts",
                "timestamp",
                "version"
            ],
            "properties": {
                "current": {
                    "description": "当前日期时间字符串",
                    "type": "string",
                    "example": "2025-01-08 10:25:20"
                },
                "docs": {
                    "description": "接口文档",
                    "allOf": [
                        {
                            "$ref": "#/definitions/routers.PingDocsScheme"
                        }
                    ]
                },
                "message": {
                    "description": "保留字段",
                    "type": "string",
                    "example": "pong"
                },
                "run_at": {
                    "description": "服务启动时的日期时间字符串",
                    "type": "string",
                    "example": "2025-01-08 10:25:20"
                },
                "run_at_ts": {
                    "description": "服务启动时的时间戳",
                    "type": "integer",
                    "example": 1736303120
                },
                "timestamp": {
                    "description": "当前时间戳",
                    "type": "integer",
                    "example": 1736303120
                },
                "version": {
                    "description": "版本号",
                    "type": "string",
                    "example": "0.1.0"
                }
            }
        },
        "xml.ConvertXMLInput": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "xml 字符串",
                    "type": "string",
                    "example": "\u003cnote\u003e\u003cto\u003evalue\u003c/to\u003e\u003c/note\u003e"
                }
            }
        },
        "xml.ConvertXMLOutput": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "json 字符串",
                    "type": "string",
                    "example": "{\"note\": {\"to\": \"value\"}}"
                }
            }
        }
    }
}